+++ code/server/app/src/main/java/server/web/param/auth/UserSession.java
@@ -1,11 +1,11 @@
 package server.web.param.auth;
 
 public class UserSession {
-    public int user_id;
-    public int session_id;
+    public long user_id;
+    public long session_id;
     public String email;
 
-    public Integer organizer_id;
+    public Long organizer_id;
     public Boolean has_analytics;
 
     public boolean admin;
