+++ code/server/app/src/main/java/server/web/root/api/OrganizerAPI.java
@@ -20,14 +20,14 @@ public class OrganizerAPI {
                 throw new ClientError.BadRequest("Account already an organizer");
         }
 
-        int organizer_id;
+        long organizer_id;
         try(var stmt = trans.namedPreparedStatement("insert into organizers values(null, false) returning id")){
             organizer_id = stmt.executeQuery().getInt("id");
         }
 
         try(var stmt = trans.namedPreparedStatement("update users set organizer_id=:organizer_id where id=:user_id")){
-            stmt.setInt(":organizer_id", organizer_id);
-            stmt.setInt(":user_id", session.user_id);
+            stmt.setLong(":organizer_id", organizer_id);
+            stmt.setLong(":user_id", session.user_id);
             if(stmt.executeUpdate()!=1)
                 throw new ClientError.BadRequest("Failed to make account an organizer");
         }
